Add course instance progress and module completions sections
- Introduced CourseInstanceProgressSection and CourseModuleCompletionsSection components to display user progress and module completion details for course instances.
- Implemented ExerciseListSection to organize exercises by chapter, enhancing user navigation and interaction.
- Created hooks for fetching course instance progress, module completions, and exercise status summaries, improving data management and component integration.
- Updated CourseInstanceExerciseStatusList to utilize new components and hooks, streamlining the course status summary page.
